v6: use less dynamic JSON parsing
The Go encoding/json package makes it easy to define types
that represent the shape of the expected data, which lets the
JSON parser do most of the heavy lifting around checking for
type compatibility.

This PR changes the code to follow this approach, arguably
making the code cleaner and  easier to follow.

Also define an `operator` type which is idiomatic in Go for representing
a known set of constants, rather than using literal numbers in the codee.