v6: remove policy factory
The `refreshPolicyFactory` type seems confusing because it's
an interface type but contains methods specific to all the various
policy types. Rename it to `refreshPolicyConfig`, remove
the visitor methods, and rename `RefreshMode.accept` to `RefreshMode.refreshPolicy` to better reflect its role in creating an
appropriate refresh policy for the refresh mode.