v6: avoid internal tests
This changes the tests to use a more end-to-end approach
by starting up a temporary HTTP server and pointing a client
to that, instead of faking out the configFetcher.

This means that the tests still remain useful when the internal
details are refactored.

Also: provide a way to optionally run the integration tests without actually
hitting the server, and make the tests run faster.

Also: use an external test helper package to make test failures nicer.