v6: more external tests
This changes the config fetcher tests to be covered by external
tests instead. The test coverage should be somewhat higher.

Also use a custom logger that emits to `t.Log` to avoid pollution
when running the tests.

Also fix an issue which meant that the most recent config wasn't
reused if the cache failed after succeeding.