v6: more benchmarks
add some more benchmarks so that we
can compare costs with v7.

Note that the `Get/with-percentage` benchmark
is exactly the same as the old `GetValue` benchmark
except that it also includes creation of the user value
as part of the benchmark.

```
name                      time/op
Get/one-of-8              314ns ± 0%
Get/less-than-with-int-8  403ns ± 2%
Get/with-percentage-8     635ns ±14%

name                      alloc/op
Get/one-of-8               360B ± 0%
Get/less-than-with-int-8   360B ± 0%
Get/with-percentage-8      456B ± 0%

name                      allocs/op
Get/one-of-8               3.00 ± 0%
Get/less-than-with-int-8   3.00 ± 0%
Get/with-percentage-8      5.00 ± 0%
```