v7: avoid copying mutex when copying configuration
`go vet` was complaining (with good reason) about the copy of the
configuration that `withFetchTime` does. Avoid the error by
using a `*sync.Map` instead of `sync.Map` directly.
Another possible solution would be to avoid putting the
fetch time inside the configuration, but it seems like that
would make the rest of the code more complex for not much gain.