v7: add tests for log output
After the PR feedback on https://github.com/configcat/go-sdk/pull/41,
I realised that a significant current issue is the lack of tests for
the expected logging output, meaning that it's easy to change
code such that the logged output changes without breaking tests.

This PR adds some tests for logs.

We can see that some improvement could be made - for example
none of the logs mention what key is currently being evaluated - but
we'll leave that for another PR.