v7: add Key method to flags
The key string is already there, so this just provides an accessor for it.

This also exposed an existing bug where the `BoolFlag.key` field wasn't initialized,
so logger messages involving bool flags would sometimes show an empty key.

Add an extra logger test to check an error message that wasn't previously tested,
and make sure that the keys returned by `GetAllKeys` are always returned in
lexical order so that the test consistently passes.