v7: better tests
In preparation for the performance improvements in #46,
this PR updates the tests to test more aspects of the implementation.

One notable change is that the type of the "expected" value in the integration
tests is no longer derived from the value received by the implementation.
Unfortunately the test data does not specifically say what type is
expected, but in practice it's always specific enough that we can
determine the type (in particular floating point numbers always have
a decimal point and booleans are always formatted as `True` or `False`.

This change exposed an issue with the percentage rule value fixup which
meant that the values weren't being updated, which is now fixed.